refactor: extract hooks for alerts and theme
