+++ src/App.tsx
@@ -8,7 +8,8 @@ import { ThemeProvider as StyledThemeProvider } from 'styled-components';
 import AlertsProvider from './contexts/alerts.context';
 import AuthProvider from './contexts/auth.provider';
 import SettingsPage from './UI/pages/settings/settings.page';
-import { ThemeProvider, useTheme } from './contexts/theme.context';
+import { ThemeProvider } from './contexts/theme.context';
+import { useTheme } from './hooks/theme.hook';
 
 const ThemedApp = () => {
   const { isDarkMode } = useTheme();
