+++ src/UI/components/navigationBar/navigation.bar.tsx
@@ -9,7 +9,7 @@ import LogoutRoundedIcon from '@mui/icons-material/LogoutRounded';
 import TuneRoundedIcon from '@mui/icons-material/TuneRounded';
 import { Fragment, useEffect } from 'react';
 import { useUsersListerDispatchContext, useUsersListerStateContext } from '../../../reducers/auth.reducer';
-import { useAlerts } from '../../../contexts/alerts.context';
+import { useAlerts } from '../../../hooks/alerts.hook';
 import FirebaseAuthManager from '../../../network/authentication/firebase.auth.manager';
 import { UserRepositoryImpl } from '../../../network/repositories/user.respository';
