+++ src/hooks/alerts.hook.ts
@@ -1,0 +1,28 @@
+import { useContext, useRef, useState } from 'react';
+import { AlertsContext, AlertType } from '../contexts/alerts.context';
+
+export const useAlerts = () => {
+  const context = useContext(AlertsContext);
+  if (!context) {
+    throw new Error('useAlerts must be used within an AlertsProvider');
+  }
+
+  const { addAlert, dismissAlert } = context;
+
+  const [alertIds, setAlertIds] = useState<string[]>([]);
+  const alertIdsRef = useRef<string[]>(alertIds);
+
+  const addAlertWithId = (alert: Omit<AlertType, 'id'>): void => {
+    const id = addAlert(alert);
+    alertIdsRef.current.push(id);
+    setAlertIds([...alertIdsRef.current]);
+  };
+
+  const clearAlerts = (): void => {
+    alertIdsRef.current.forEach(dismissAlert);
+    alertIdsRef.current = [];
+    setAlertIds([]);
+  };
+
+  return { addAlert: addAlertWithId, clearAlerts };
+};
